Uber Clone Script
Every taxi business needs an optimized and customized Uber clone app via which the business runs smoothly. Many aspiring entrepreneurs are making use of the clone apps to usher in their on-demand taxi business. Any taxi business owner can make use of the high quality service provided by the experts at Uber Like App to build their online apps. Since it is the most productive tool, our developers and business experts come together to ensure that the app is highly user-friendly and feature-rich. The apps we offer are compatible with multiple platforms and it can be launched in the market in a short period. Our premium features include:

•	Hassle-free login 
•	Saved addresses
•	Call masking for security
•	God’s eye view
•	Acceptance/cancellation of rides
•	Effortless vehicle management
•	SOS button for emergencies
•	Fix prices effortlessly